requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
protobuf==3.12.2
pytesseract==0.3.4
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
tqdm==4.47.0
Werkzeug==1.0.1
wrapt==1.12.1
